# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "嘉阅圈私域社群平台后端服务，提供内容社区、互动打卡、小组交流、私信通知、知识课程等核心能力。\n统一响应格式：{\"success\":bool,\"data\":any,\"message\":string,\"error\":{\"code\":string,\"message\":string}}",
        "title": "嘉阅圈后端 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3100",
    "basePath": "/api",
    "paths": {
...
        "/checkin": {
            "get": {
                "description": "公开接口。分页查询打卡记录，可按用户 ID 与圈子 ID 过滤。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "打卡"
                ],
                "summary": "打卡记录列表",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户 ID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "圈子 ID",
                        "name": "circleId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，默认 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量，默认 20",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "需登录。兼容旧客户端入口，委托 PostService 创建打卡挑战帖子，并在同一事务计算连续天数；同一用户同一圈子每日仅可打卡一次。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "打卡"
                ],
                "summary": "创建打卡",
                "parameters": [
                    {
                        "description": "打卡请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式错误或圈子 ID 不合法",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "今日已打卡",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "打卡记录创建失败",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "yjedu-reading-club-server_internal_pkg_response.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "yjedu-reading-club-server_internal_pkg_response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \u003ctoken\u003e\"，token 通过 /api/auth/login/phone 或 /api/auth/weapp/login 获取。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}